+++ src/scripts/goc_challenge1_huristic_cli.jl
@@ -1,6 +1,6 @@
 #!/usr/bin/env julia
 
-include("goc_challenge1_huristic.jl")
 include("scopf-main.jl")
+include("goc_challenge1_huristic.jl")
 
 scopf_main(parse_scopf_commandline())