+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerModelsSecurityConstrained.jl Change Log
 ============================================
 
 ### Staged
+- Add use Gurobi option to cli interface
 - Minor fixes to cli interface in scripts
 
 ### v0.4.0